+++ src/test/java/org/nrg/containers/rest/CommandRestApiTest.java
@@ -60,7 +60,7 @@ import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
-import static org.nrg.containers.services.impl.ContainerConfigServiceImpl.WRAPPER_CONFIG_PATH_TEMPLATE;
+import static org.nrg.containers.services.ContainerConfigService.WRAPPER_CONFIG_PATH_TEMPLATE;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.authentication;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.csrf;
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.testSecurityContext;